// src/components/Tabs/UserModel/UserModel.interface.ts
export enum Online {
	ONLINE = 'ONLINE',
	OFFLINE = 'OFFLINE',
}

export enum OnlineView {
	ONLINE = 'В сети',
	OFFLINE = 'Не в сети',
}

export enum sortTitles {
	CHECK = 'CHECK',
	ID = 'ID',
	USERNAME = 'USERNAME',
	EMAIL = 'EMAIL',
	ROLE = 'ROLE',
	ONLINE = 'ONLINE',
	AVATAR = 'AVATAR',
	PHONE = 'PHONE',
	ORDERS = 'ORDERS',
	DATE_UPDATED = 'DATE_UPDATED',
	DATE_CREATED = 'DATE_CREATED',
}

export enum sortTitlesView {
	CHECK = ' ',
	ID = '№',
	USERNAME = 'Имя пользователя',
	EMAIL = 'Почта',
	ROLE = 'Роль',
	ONLINE = 'Онлайн',
	AVATAR = 'Аватарка',
	PHONE = '№ телефона',
	ORDERS = 'Заказы',
	DATE_UPDATED = 'Дата обновления',
	DATE_CREATED = 'Дата регистрации',
}

export enum sortTitlesSend {
	USERNAME = 'username',
	EMAIL = 'email',
	PHONE = 'phone',
}